+++ Realm.Shared/linq/RealmResults.cs
@@ -20,26 +20,44 @@ namespace Realms
         public Type ElementType => typeof (T);
         public Expression Expression { get; } = null; // null if _allRecords
         private readonly RealmResultsProvider _provider = null;  // null if _allRecords
-        private bool _allRecords = false;
+        private readonly bool _allRecords = false;
         private readonly Realm _realm;
+
+        internal ResultsHandle ResultsHandle => _resultsHandle ?? (_resultsHandle = CreateResultsHandle()); 
         private ResultsHandle _resultsHandle = null;
 
         public IQueryProvider Provider => _provider;
 
-        internal RealmResults(RealmResultsProvider queryProvider, Expression expression) 
-        {
-            this._provider = queryProvider;
-            _realm = _provider?._realm ?? null;
-            this.Expression = expression;
-        }
-
-        internal RealmResults(Realm realm, bool createdByAll=false)
+        internal RealmResults(Realm realm, RealmResultsProvider realmResultsProvider, Expression expression,
+            bool createdByAll)
         {
             _realm = realm;
+            _provider = realmResultsProvider;
+            Expression = expression ?? Expression.Constant(this);
             _allRecords = createdByAll;
-            _provider = new RealmResultsProvider(realm);
-            this.Expression = Expression.Constant (this);
 
+        }
+
+        internal RealmResults(Realm realm, bool createdByAll)
+            : this(realm, new RealmResultsProvider(realm), null, createdByAll)
+        {
+        }
+
+        private ResultsHandle CreateResultsHandle()
+        {
+            var retType = typeof (T);
+            if (_allRecords)
+            {
+                return _realm.MakeResultsForTable(retType);
+            }
+            else
+            {
+                // do all the LINQ expression evaluation to build a query
+                var qv = _provider.MakeVisitor(retType);
+                qv.Visit(Expression);
+                var queryHandle = qv._coreQueryHandle; // grab out the built query definition
+                return _realm.MakeResultsForQuery(retType, queryHandle);
+            }
         }
 
         /// <summary>
@@ -48,27 +66,10 @@ namespace Realms
         /// <returns>An IEnumerator which will iterate through found Realm persistent objects.</returns>
         public IEnumerator<T> GetEnumerator()
         {
-            var retType = typeof(T);
-            if (_resultsHandle == null)  // first call
-            {
-                if (_allRecords)
-                {
-                    _resultsHandle = _realm.MakeResultsForTable(retType);
-                }
-                else
-                {
-                    // do all the LINQ expression evaluation to build a query
-                    var qv = _provider.MakeVisitor(retType);
-                    qv.Visit(this.Expression);
-                    var queryHandle = qv._coreQueryHandle;  // grab out the built query definition
-                    _resultsHandle = _realm.MakeResultsForQuery(retType, queryHandle);
-                }
-            }
-            return new RealmResultsEnumerator<T>(_realm, _resultsHandle);
+            return new RealmResultsEnumerator<T>(_realm, ResultsHandle);
         }
 
         IEnumerator IEnumerable.GetEnumerator()
-
         {
             return (IEnumerator)GetEnumerator();  // using our class generic type, just redirect the legacy get
         }